# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Apex Hanseikai API",
    "version": "3.0.0",
    "description": "Recordings, notes (hansei), principles and Watch Party. There is no authentication, so keep the server on a trusted network. Setup and the principle Markdown are in the guides: https://hrs23.github.io/apex-hanseikai/docs/en/reference.html. For LLMs: https://hrs23.github.io/apex-hanseikai/docs/llms.txt"
  },
  "paths": {
...
    "/api/recordings": {
      "get": {
        "summary": "List recordings, newest first",
        "tags": [
          "media"
        ],
        "responses": {
          "200": {
            "description": "Newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recordings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Recording"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid day"
          }
        },
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "description": "Filter by day and prioritize generation for its recordings.",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "description": "Listing also queues any recording whose audio, match ends or thumbnails are missing."
      }
    },
    "/api/derived": {
      "post": {
        "summary": "Make or remake one derived file of a recording",
        "tags": [
          "media"
        ],
        "description": "`low` makes the 720p version, which is never made on its own. `audio`, `markers` and `thumbs` remake the player audio, match ends and thumbnails. The work is queued and waits while a recording is in progress; follow it in `GET /api/recordings`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recording",
                  "stage"
                ],
                "properties": {
                  "recording": {
                    "type": "string",
                    "example": "2026-10-03 22-54-36.mp4"
                  },
                  "stage": {
                    "type": "string",
                    "enum": [
                      "audio",
                      "markers",
                      "thumbs",
                      "low"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued"
          },
          "400": {
            "description": "Unknown recording or stage"
          }
        }
      }
    },
    "/api/hansei": {
      "get": {
        "summary": "List notes (filter by recording)",
        "tags": [
          "hansei"
        ],
        "parameters": [
          {
            "name": "recording",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact recording filename. Omit to list all",
            "example": "2026-10-03 22-54-36.mp4"
          }
        ],
        "responses": {
          "200": {
            "description": "Ascending by time (entries without a time last)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hansei": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Hansei"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          }
        },
        "description": "Notes sorted by time. Pass `recording` to get one recording's notes."
      },
      "post": {
        "summary": "Add a note",
        "tags": [
          "hansei"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HanseiInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The added 反省",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hansei"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "description": "Body exceeds 128 KiB"
          }
        },
        "description": "`at` is the absolute UTC time of the moment, or null for the whole recording. The body is trimmed and limited to 500 characters. Other viewers see it right away over `/sync/stream`."
      }
    },
...
    "/api/log": {
      "post": {
        "summary": "Report an error from the UI (logged as client_error)",
        "tags": [
          "system"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "stack": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "detail": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged"
          },
          "400": {
            "description": "Invalid body"
          },
          "429": {
            "description": "More than 60 reports in a minute"
          }
        },
        "description": "Internal: used by the web UI to report errors. Not meant for other clients. Rate limited (429). Entries appear in the server log."
      }
    },
...
  },
  "components": {
    "responses": {
      "Invalid": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Hansei": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Absolute time (UTC), recording `startedAt` + position. null for a note about the whole recording"
          },
          "body": {
            "type": "string",
            "description": "Note text, up to 500 characters"
          },
          "created_at": {
            "type": "string",
            "description": "When the note was added"
          },
          "recording": {
            "type": "string",
            "description": "Recording file name"
          }
        }
      },
      "HanseiInput": {
        "type": "object",
        "required": [
          "recording",
          "at",
          "body"
        ],
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "recording": {
            "type": "string",
            "example": "2026-10-03 22-54-36.mp4"
          }
        }
      },
      "Recording": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "2026-10-03 22-54-36.mp4"
          },
          "state": {
            "type": "string",
            "enum": [
              "recording",
              "processing",
              "ready"
            ],
            "description": "recording: still being written. processing: audio, match ends or thumbnails are still being generated. ready: nothing left to do"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation time stored in the file, or the time in the file name when there is none"
          },
          "duration": {
            "type": "number",
            "description": "Seconds. 0 while unknown"
          },
          "players": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Names of audio tracks 2 and up, one per player, in cell order. Empty for a single-track recording or until the audio is extracted"
          },
          "thumbnails": {
            "type": "boolean"
          },
          "matchEnds": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Match end positions in seconds"
          },
          "height": {
            "type": "integer",
            "description": "Picture height in pixels. 0 while unknown"
          },
          "low": {
            "type": "string",
            "nullable": true,
            "enum": [
              "ready",
              "processing",
              null
            ],
            "description": "The 720p version: ready, being made, or null when nobody asked for it"
          }
        }
      },
...
    }
  },
...
}